Fixed small mistake.

+++ kernel/src/drivers/pci/impl/pci.c
@@ -186,31 +186,31 @@ static uint16_t get_bar5_high(uint8_t bus, uint8_t slot, uint8_t func) {
 }
 
 
-uint8_t pci_get_bar0(uint8_t bus, uint8_t slot, uint8_t func) {
+uint32_t pci_get_bar0(uint8_t bus, uint8_t slot, uint8_t func) {
     return get_bar0_low(bus, slot, func) | get_bar0_high(bus, slot, func);
 }
 
 
-uint8_t pci_get_bar1(uint8_t bus, uint8_t slot, uint8_t func) {
+uint32_t pci_get_bar1(uint8_t bus, uint8_t slot, uint8_t func) {
     return get_bar1_low(bus, slot, func) | get_bar1_high(bus, slot, func);
 }
 
 
-uint8_t pci_get_bar2(uint8_t bus, uint8_t slot, uint8_t func) {
+uint32_t pci_get_bar2(uint8_t bus, uint8_t slot, uint8_t func) {
     return get_bar2_low(bus, slot, func) | get_bar2_high(bus, slot, func);
 }
 
 
-uint8_t pci_get_bar3(uint8_t bus, uint8_t slot, uint8_t func) {
+uint32_t pci_get_bar3(uint8_t bus, uint8_t slot, uint8_t func) {
     return get_bar3_low(bus, slot, func) | get_bar3_high(bus, slot, func);
 }
 
 
-uint8_t pci_get_bar4(uint8_t bus, uint8_t slot, uint8_t func) {
+uint32_t pci_get_bar4(uint8_t bus, uint8_t slot, uint8_t func) {
     return get_bar4_low(bus, slot, func) | get_bar4_high(bus, slot, func);
 }
 
 
-uint8_t pci_get_bar5(uint8_t bus, uint8_t slot, uint8_t func) {
+uint32_t pci_get_bar5(uint8_t bus, uint8_t slot, uint8_t func) {
     return get_bar5_low(bus, slot, func) | get_bar5_high(bus, slot, func);
 }
